pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-cra-scheduler-env"
version = "0.1.0"
description = "CRA Site Visit Scheduler Environment for OpenEnv — multi-CRA routing with time windows"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "ortools>=9.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "cra_scheduler_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["cra_scheduler_env", "cra_scheduler_env.server"]
package-dir = { "cra_scheduler_env" = ".", "cra_scheduler_env.server" = "server" }
